+++ app/src/main/java/com/haystack/photosearch/ui/gallery/PhotoGalleryViewModel.kt
@@ -1,4 +1,4 @@
-package com.haystack.photosearch.ui.main
+package com.haystack.photosearch.ui.gallery
 
 import android.util.Log
 import androidx.lifecycle.ViewModel
